+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### v1.2.0
+* add `access list` methods
+* fix revoke token
 ### v1.1.3
 * update to new log
 * migrate to new db client